token: extract Nonterminal::to_tokenstream to parser.
